Moving more common code
