+++ simkratos/DEM/kratos_DEM_wrapper.py
@@ -34,6 +34,8 @@ class DEMWrapper(KratosWrapper):
         self.time = 0
         self.step = 0
         self.substeps = 0
+        self.particle_type = "SphericParticle3D"
+        self.condition_type = "RigidFace3D3N"
 
         # The dictionary defines the relation between CUBA and
         # kratos variables
@@ -153,96 +155,6 @@ class DEMWrapper(KratosWrapper):
 
                 dst.update_particles([particle])
 
-    def exportKratosElements(self, src, dst, group):
-        """ Parses all kratos elements to simphony cells
-
-        Iterates over all nodes in the kratos mesh (src) and
-        converts them to simphony cells (dst). While doing this operation
-        any point that has not currently mapped will have his uuid
-        added in the 'id_map' of the weapper
-
-        """
-
-        for element in src.GetElements(group):
-
-            element_uid = None
-
-            data = {}
-
-            if element.Id not in self.id_to_uuid_element_map:
-
-                point_list = [
-                    self.id_to_uuid_node_map[pointl.Id]
-                    for pointl in element.GetNodes()
-                ]
-
-                cell = SCell(
-                    points=point_list,
-                    data=DataContainer(data),
-                    uid=element_uid
-                )
-
-                cid = dst.add([cell])
-
-                self.id_to_uuid_element_map[element.Id] = cid[0]
-
-            else:
-
-                # No data is stored in the element yet
-
-                pass
-
-    def importKratosNodes(self, src, dst, group):
-        """ Parses all simphony points to kratos nodes
-
-        Iterates over all points in the simphony mesh (src) and
-        converts them to kratos nodes (dst).
-        While doing this operation any point/node pair that has not
-        currently mapped will have  his uuid added in the 'id_map'
-        of the wrapper
-
-        """
-
-        # Add the points in case they don't exist and update their value in
-        # case they do.
-        for point in src.iter(item_type=CUBA.POINT):
-
-            if point.uid not in self.uuid_to_id_node_map.keys():
-                self.uuid_to_id_node_map.update(
-                    {point.uid: self.free_id}
-                )
-
-                self.free_id += 1
-
-                node_id = self.uuid_to_id_node_map[point.uid]
-
-                data = point.data
-
-                node = dst.CreateNewNode(
-                    node_id,
-                    point.coordinates[0],
-                    point.coordinates[1],
-                    point.coordinates[2])
-
-                self.setNodalData(data, node, dst.Name)
-
-                self.id_to_ref_node[node_id] = node
-
-            else:
-
-                node = self.id_to_ref_node[self.uuid_to_id_node_map[point.uid]]
-                data = point.data
-                self.setNodalData(data, node, dst.Name)
-
-        # If they belong to a different group, add them
-        if group != 0:
-            nodes = KRTS.NodesArray()
-            for point in src.iter(item_type=CUBA.POINT):
-                nodes.append(
-                    dst.Nodes[self.uuid_to_id_node_map[point.uid]]
-                )
-            dst.SetNodes(nodes, group)
-
     def importKratosParticles(self, src, dst, group):
         """ Parses all simphony particles to kratos particle elements
 
@@ -315,80 +227,6 @@ class DEMWrapper(KratosWrapper):
                 )
             dst.SetNodes(nodes, group)
             dst.SetElements(elements, group)
-
-    def importKratosElements(self, src, dst, group):
-        """ Parses all simphony cells to kratos elements
-
-        Iterates over all cells in the simphony mesh (src) and
-        converts them to kratos SphericContinuumParticle3D elements (dst).
-        While doing this operation any point/node pair that has not
-        currently mapped will have  his uuid added in the 'id_map'
-        of the wrapper
-
-        """
-
-        for element in src.iter(item_type=CUBA.CELL):
-
-            if element.uid not in self.uuid_to_id_element_map.keys():
-                self.uuid_to_id_element_map.update(
-                    {element.uid: self.free_id}
-                )
-
-                self.free_id += 1
-
-                element_id = self.uuid_to_id_element_map[element.uid]
-
-                dst.CreateNewElement(
-                    "SphericParticle3D",
-                    element_id,
-                    [self.uuid_to_id_node_map[p] for p in element.points],
-                    self.element_properties)
-
-        # If they belong to a different group, add them
-        if group != 0:
-            elements = KRTS.ElementsArray()
-            for elem in src.iter(item_type=CUBA.CELL):
-                elements.append(
-                    dst.Elements[self.uuid_to_id_element_map[elem.uid]]
-                )
-            dst.SetElements(elements, group)
-
-    def importKratosConditions(self, src, dst, group):
-        """ Parses all simphony faces to kratos conditions
-
-        Iterates over all faces in the simphony mesh (src) and
-        converts them to kratos RigidFace3D3N conditions (dst).
-        While doing this operation any point/node pair that has not
-        currently mapped will have  his uuid added in the 'id_map'
-        of the wrapper
-
-        """
-
-        for condition in src.iter(item_type=CUBA.FACE):
-
-            if condition.uid not in self.uuid_to_id_condition_map.keys():
-                self.uuid_to_id_condition_map.update(
-                    {condition.uid: self.free_id}
-                )
-
-                self.free_id += 1
-
-                condition_id = self.uuid_to_id_condition_map[condition.uid]
-
-                dst.CreateNewCondition(
-                    "RigidFace3D3N",
-                    condition_id,
-                    [self.uuid_to_id_node_map[p] for p in condition.points],
-                    self.condition_properties)
-
-        # If they belong to a different group, add them
-        if group != 0:
-            conditions = KRTS.ConditionsArray()
-            for cnd in src.iter(item_type=CUBA.FACE):
-                conditions.append(
-                    dst.Conditions[self.uuid_to_id_condition_map[cnd.uid]]
-                )
-            dst.SetConditions(conditions, group)
 
     def _setMeshData(self):
         " This probably needs to be done throug configuration"
