+++ simkratos/kratosWrapper.py
@@ -24,6 +24,9 @@ from simphony.cuds.mesh import Point as SPoint
 from simphony.cuds.mesh import Face as SFace
 from simphony.cuds.mesh import Cell as SCell
 
+# Kratos Imports
+import KratosMultiphysics as KRTS
+
 
 class KratosWrapper(ABCModelingEngine):
     """ BaseClass for Kratos based wrappers
@@ -62,7 +65,7 @@ class KratosWrapper(ABCModelingEngine):
         self.meshes = []
         self.lattices = []
 
-        # Id's stuff
+        # Entity count
         self.free_id = 1
 
         # Simphony
@@ -372,7 +375,7 @@ class KratosWrapper(ABCModelingEngine):
 
             data = {}
 
-            self.getNodalData(data, node)
+            self.getNodalData(data, node, src.Name)
 
             point_uid = None
 
@@ -474,6 +477,131 @@ class KratosWrapper(ABCModelingEngine):
                 # No data is stored in the element yet
 
                 pass
+
+    def importKratosNodes(self, src, dst, group):
+        """ Parses all simphony points to kratos nodes
+
+        Iterates over all points in the simphony mesh (src) and
+        converts them to kratos nodes (dst).
+        While doing this operation any point/node pair that has not
+        currently mapped will have  his uuid added in the 'id_map'
+        of the wrapper
+
+        """
+
+        # Add the points in case they don't exist and update their value in
+        # case they do.
+        for point in src.iter(item_type=CUBA.POINT):
+
+            if point.uid not in self.uuid_to_id_node_map.keys():
+                self.uuid_to_id_node_map.update(
+                    {point.uid: self.free_id}
+                )
+
+                self.free_id += 1
+
+                node_id = self.uuid_to_id_node_map[point.uid]
+
+                data = point.data
+
+                node = dst.CreateNewNode(
+                    node_id,
+                    point.coordinates[0],
+                    point.coordinates[1],
+                    point.coordinates[2])
+
+                self.setNodalData(data, node, dst.Name)
+
+                self.id_to_ref_node[node_id] = node
+
+            else:
+
+                node = self.id_to_ref_node[self.uuid_to_id_node_map[point.uid]]
+                data = point.data
+                self.setNodalData(data, node, dst.Name)
+
+        # If they belong to a different group, add them
+        if group != 0:
+            nodes = KRTS.NodesArray()
+            for point in src.iter(item_type=CUBA.POINT):
+                nodes.append(
+                    dst.Nodes[self.uuid_to_id_node_map[point.uid]]
+                )
+            dst.SetNodes(nodes, group)
+
+    def importKratosElements(self, src, dst, group, element_type):
+        """ Parses all simphony cells to kratos elements
+
+        Iterates over all cells in the simphony mesh (src) and
+        converts them to kratos SphericContinuumParticle3D elements (dst).
+        While doing this operation any point/node pair that has not
+        currently mapped will have  his uuid added in the 'id_map'
+        of the wrapper
+
+        """
+
+        for element in src.iter(item_type=CUBA.CELL):
+
+            if element.uid not in self.uuid_to_id_element_map.keys():
+                self.uuid_to_id_element_map.update(
+                    {element.uid: self.free_id}
+                )
+
+                self.free_id += 1
+
+                element_id = self.uuid_to_id_element_map[element.uid]
+
+                dst.CreateNewElement(
+                    element_type,
+                    element_id,
+                    [self.uuid_to_id_node_map[p] for p in element.points],
+                    self.element_properties)
+
+        # If they belong to a different group, add them
+        if group != 0:
+            elements = KRTS.ElementsArray()
+            for elem in src.iter(item_type=CUBA.CELL):
+                elements.append(
+                    dst.Elements[self.uuid_to_id_element_map[elem.uid]]
+                )
+            dst.SetElements(elements, group)
+
+    def importKratosConditions(self, src, dst, group, condition_type):
+        """ Parses all simphony faces to kratos conditions
+
+        Iterates over all faces in the simphony mesh (src) and
+        converts them to kratos RigidFace3D3N conditions (dst).
+        While doing this operation any point/node pair that has not
+        currently mapped will have  his uuid added in the 'id_map'
+        of the wrapper
+
+        """
+
+        for condition in src.iter(item_type=CUBA.FACE):
+
+            if condition.uid not in self.uuid_to_id_condition_map.keys():
+                self.uuid_to_id_condition_map.update(
+                    {condition.uid: self.free_id}
+                )
+
+                self.free_id += 1
+
+                condition_id = self.uuid_to_id_condition_map[condition.uid]
+
+                dst.CreateNewCondition(
+                    condition_type,
+                    condition_id,
+                    [self.uuid_to_id_node_map[p] for p in condition.points],
+                    self.condition_properties)
+
+        # If they belong to a different group, add them
+        if group != 0:
+            conditions = KRTS.ConditionsArray()
+            for cnd in src.iter(item_type=CUBA.FACE):
+                conditions.append(
+                    dst.Conditions[self.uuid_to_id_condition_map[cnd.uid]]
+                )
+            dst.SetConditions(conditions, group)
 
     def run(self, mesh):
         pass
